// src/fengshui/types.ts
export interface Position3D {
  x: number;
  y: number;
  z: number;
}

export type Direction = "乾" | "坤" | "震" | "巽" | "坎" | "离" | "艮" | "兑";

export type Mountain =
  | "壬"
  | "子"
  | "癸"
  | "丑"
  | "艮"
  | "寅"
  | "甲"
  | "卯"
  | "乙"
  | "辰"
  | "巽"
  | "巳"
  | "丙"
  | "午"
  | "丁"
  | "未"
  | "坤"
  | "申"
  | "庚"
  | "酉"
  | "辛"
  | "戌"
  | "乾"
  | "亥";